contra-costa-knowledge-bot/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Data handling
pydantic==2.9.2
python-dotenv==1.0.1

# RAG stack
openai==1.44.0
faiss-cpu==1.8.0.post1

# Utilities
numpy==1.26.4
pandas==2.2.3
requests==2.32.3
